+++ app/src/assets/images/backgrounds/landing-bg-light.png
