+++ app/src/assets/images/icons/logo-alt.png
